truck_routing/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.2.23
django-filter==23.5
djangorestframework==3.13.0
djangorestframework-gis==1.0
idna==3.10
psycopg2==2.9.10
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
